update loginpage and registerpage
